Updated index.html
You had some errors that I fixed. Its common practice to have the <!DOCTYPE html> at the top of the file.
You had some html code in the "container div" before the <html> tag, so that would have thrown an error. And then it is easier to find errors and problems if you space things out neatly. So I did some spacing changes for you.
<3